feature: add default 404 page
