+++ src/pages/404/index.tsx
@@ -1,0 +1,25 @@
+import { memo, useRef } from 'react';
+import classnames from 'classnames';
+
+import styles from './index.module.scss';
+
+import Head from '../../components/Head/Head';
+
+type Props = {
+  className: string;
+};
+
+function NotFound({ className }: Props) {
+  const containerRef = useRef<HTMLElement>(null);
+
+  return (
+    <main className={classnames(styles.NotFound, className)} ref={containerRef}>
+      <Head title="NotFound" />
+      <h1 className={styles.title}>NotFound Page</h1>
+    </main>
+  );
+}
+
+NotFound.defaultProps = {};
+
+export default memo(NotFound);
